dev_requirements.txt
django-nose
coverage
django-coverage
ipdb
flake8
fabric
factory-boy>=2.0.2
mock
Pillow>=2.1.0
django-libs>=1.42.3
South